task5 red, work in progres
